fix: move model namespaces around
