+++ src/ConfigReader.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Sigwin\Ariadne;
 
-use Sigwin\Ariadne\Model\AdrianeConfig;
+use Sigwin\Ariadne\Model\Config\AdrianeConfig;
 
 interface ConfigReader
 {
