+++ src/Model/Collection/TemplateCollection.php
@@ -11,7 +11,9 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Sigwin\Ariadne\Model;
+namespace Sigwin\Ariadne\Model\Collection;
+
+use Sigwin\Ariadne\Model\Template;
 
 /**
  * @implements \IteratorAggregate<Template>
