+++ src/Model/Config/ProfileClientConfig.php
@@ -11,7 +11,11 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Sigwin\Ariadne\Model;
+namespace Sigwin\Ariadne\Model\Config;
+
+use Sigwin\Ariadne\Model\TProfileClient;
+use Sigwin\Ariadne\Model\TProfileClientAuth;
+use Sigwin\Ariadne\Model\TProfileClientOptions;
 
 /**
  * @psalm-type TProfileClientAuth = array{type: string, token: string}
