+++ src/Model/Config/ProfileConfig.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Sigwin\Ariadne\Model;
+namespace Sigwin\Ariadne\Model\Config;
 
 /**
  * @psalm-import-type TProfileTemplate from ProfileTemplateConfig
