+++ src/ProfileFactory.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Sigwin\Ariadne;
 
-use Sigwin\Ariadne\Model\ProfileConfig;
+use Sigwin\Ariadne\Model\Config\ProfileConfig;
 
 interface ProfileFactory
 {
