+++ tests/Bridge/Gitlab/GitlabProfileTest.php
@@ -19,9 +19,9 @@ use Psr\Cache\CacheItemPoolInterface;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestInterface;
 use Sigwin\Ariadne\Bridge\Gitlab\GitlabProfile;
-use Sigwin\Ariadne\Model\ProfileConfig;
-use Sigwin\Ariadne\Model\RepositoryCollection;
-use Sigwin\Ariadne\Model\RepositoryTargetConfig;
+use Sigwin\Ariadne\Model\Collection\RepositoryCollection;
+use Sigwin\Ariadne\Model\Config\ProfileConfig;
+use Sigwin\Ariadne\Model\Config\RepositoryTargetConfig;
 use Sigwin\Ariadne\Model\Template;
 use Sigwin\Ariadne\ProfileTemplateFactory;
 
@@ -30,15 +30,15 @@ use Sigwin\Ariadne\ProfileTemplateFactory;
  *
  * @covers \Sigwin\Ariadne\Bridge\Gitlab\GitlabProfile
  *
- * @uses \Sigwin\Ariadne\Model\ProfileClientConfig
- * @uses \Sigwin\Ariadne\Model\ProfileConfig
+ * @uses \Sigwin\Ariadne\Model\Collection\RepositoryCollection
+ * @uses \Sigwin\Ariadne\Model\Collection\TemplateCollection
+ * @uses \Sigwin\Ariadne\Model\Config\ProfileClientConfig
+ * @uses \Sigwin\Ariadne\Model\Config\ProfileConfig
+ * @uses \Sigwin\Ariadne\Model\Config\ProfileTemplateConfig
+ * @uses \Sigwin\Ariadne\Model\Config\RepositoryTargetConfig
  * @uses \Sigwin\Ariadne\Model\ProfileSummary
- * @uses \Sigwin\Ariadne\Model\ProfileTemplateConfig
  * @uses \Sigwin\Ariadne\Model\ProfileUser
- * @uses \Sigwin\Ariadne\Model\RepositoryCollection
- * @uses \Sigwin\Ariadne\Model\RepositoryTargetConfig
  * @uses \Sigwin\Ariadne\Model\Template
- * @uses \Sigwin\Ariadne\Model\TemplateCollection
  *
  * @small
  */
